refactor: App 클래스에서 도메인 로직 분리 & AppTest 제거
