+++ src/App.js
@@ -1,12 +1,10 @@
 // @ts-check
 
 const User = require('./User.js');
-const Lotto = require('./Lotto.js');
 const LottoManager = require('./LottoManager.js');
 const Utils = require('./utils/Utils.js');
 const Printer = require('./utils/Printer.js');
 const { message } = require('./utils/messages.js');
-const { lotto } = require('./utils/const.js');
 
 class App {
   /** @type {User} */
@@ -24,7 +22,7 @@ class App {
     Utils.readLine(message.INPUT_AMOUNT_MESSAGE, (inputAmount) => {
       this.#user.setAmount(inputAmount);
       const amount = this.#user.getAmount();
-      const numbersList = this.generateNumbersList(amount);
+      const numbersList = this.#lottoManager.generateNumbersList(amount);
 
       this.#user.setNumbersList(numbersList);
       const userNumbersList = this.#user.getNumbersList();
@@ -32,38 +30,6 @@ class App {
 
       this.#askLottoNumbers();
     });
-  }
-
-  /**
-   *
-   * @param {number} amount
-   * @returns {number[][]}
-   */
-  generateNumbersList(amount) {
-    const list = [];
-
-    for (let i = 0; i < amount / lotto.PRICE; i++) {
-      const randomNumbers = Utils.getRandomNumbers(
-        lotto.MIN_NUMBER,
-        lotto.MAX_NUMBER,
-        lotto.MAX_COUNT
-      );
-
-      const lottoNumbers = this.publishLotto(randomNumbers);
-      list.push(lottoNumbers);
-    }
-
-    return list;
-  }
-
-  /**
-   *
-   * @param {number[]} numbers
-   * @returns {number[]}
-   */
-  publishLotto(numbers) {
-    const lotto = new Lotto(numbers);
-    return lotto.getNumbers();
   }
 
   #askLottoNumbers() {
@@ -78,33 +44,22 @@ class App {
   }
 
   #askBonusNumber() {
-    Utils.readLine(
-      `\n${message.INPUT_BONUS_NUMBER_MESSAGE}\n`,
-      (inputBonus) => {
-        this.#lottoManager.setBonusNumber(inputBonus);
-        const amount = this.#user.getAmount();
-        const userNumbersList = this.#user.getNumbersList();
+    Utils.readLine(message.INPUT_BONUS_NUMBER_MESSAGE, (inputBonus) => {
+      this.#lottoManager.setBonusNumber(inputBonus);
+      const amount = this.#user.getAmount();
+      const userNumbersList = this.#user.getNumbersList();
 
-        const statistics = this.#generateStatistics(amount, userNumbersList);
-        Printer.printStatistics(statistics.prizes, statistics.revenue);
-
-        Utils.close();
-      }
-    );
-  }
-
-  /**
-   *
-   * @param {number} amount
-   * @param {number[][]} userNumbersList
-   * @returns {{prizes:{first: number, second: number, third:number, fourth:number, fifth:number}, revenue: string}}
-   */
-  #generateStatistics(amount, userNumbersList) {
-    const prizes = this.#lottoManager.getPrizes(userNumbersList);
-    const revenue = this.#lottoManager.calculateRevenue(prizes, amount);
-
-    return { prizes, revenue };
+      const statistics = this.#lottoManager.generateStatistics(
+        amount,
+        userNumbersList
+      );
+      Printer.printStatistics(statistics.prizes, statistics.revenue);
+      Utils.close();
+    });
   }
 }
+
+const app = new App();
+app.play();
 
 module.exports = App;
